Start bringing this up to date in Remake/
There's a desktop (LWJGL3) and an HTML target for the new version, but
none of it will compile yet. The project is configured correctly, at the
very least for LWJGL3, since I ran it there before pasting in all the
old AssaultFish code. Some of said code has had some updates, but the
main old AssaultFish.java code has a ton of errors to fix.